Fix fd.o #10174: make it possible to return multiple values with no signature. More specifically: when a service method with no signature synchronously returns a tuple that is not a Struct, interpret it as a multi-valued return, rather than as a structure.
This is a common Python idiom, and returning a struct makes little sense
anyway when D-Bus lets you return multiple values.

Returned lists are still interpreted as arrays - returning an array is
entirely sensible, and indeed likely to be common.

Async service methods are unaffected (there is no ambiguity), and it's still
possible to return a structure by returning a dbus.Struct with appropriate
contents.

https://bugs.freedesktop.org/show_bug.cgi?id=10174