Omit the remote traceback from certain D-Bus errors
Specifically, DBusException and its subclasses no longer have the remote
traceback by default (although subclasses can turn it back on again
by setting include_traceback = True, and the various "programmer error"
subclasses of DBusException do have this set).

Hopefully this will stop people thinking it's a dbus-python or
telepathy-python bug when a D-Bus API like Telepathy deliberately raises an
error (and so dbus-python or telepathy-python is visible in the traceback).